Create Hospital Management System
I was tasked for a school project to develop a Hospital Management System to efficiently manage patient records at a local medical facility. The system should allow hospital staff to search for patient for patient records based on specific criteria and sort the records for easy access. To achieve this, I has to implement a Java GUI application with a JList componment for displaying and interacting with patient records.